third_party/blink/web_tests/platform/mac-mac10.13/virtual/storage-access-api/external/wpt/storage-access-api/storageAccess.testdriver.sub-expected.txt

This is a testharness.js-based test.
PASS Set up storage access rules
PASS [third-party-allowed-on-first-party-site] Cookie access is allowed: true
FAIL [third-party-blocked-all] Cookie access is allowed: false assert_equals: expected "" but got "hello2=world2"
FAIL [third-party-blocked-on-first-party-site] Cookie access is allowed: false assert_equals: expected "" but got "hello1=world1"
PASS [same-site] Cookie access is allowed: true
Harness: the test ran to completion.
